refactor seperate components to own files
